Refactor client.commands module
