+++ bcbiovm/client/subcommands/docker.py
@@ -1,0 +1,121 @@
+"""Utilities to help with develping using bcbion inside of docker."""
+
+from bcbiovm.client import base
+from bcbiovm.common import constant
+from bcbiovm.docker import devel as docker_devel
+
+
+class Build(base.BaseCommand):
+
+    """Build docker image and export to S3."""
+
+    def setup(self):
+        """Extend the parser configuration in order to expose this command."""
+        parser = self._main_parser.add_parser(
+            "dockerbuild",
+            help="Build docker image and export to S3")
+        parser.add_argument(
+            "-b", "--bucket", default="bcbio_nextgen",
+            help="S3 bucket to upload the gzipped docker image to")
+        parser.add_argument(
+            "-t", "--buildtype", default="full", choices=["full", "code"],
+            help=("Type of docker build to do. full is all code and third"
+                  " party tools. code is only bcbio-nextgen code."))
+        parser.add_argument(
+            "-d", "--rundir", default="/tmp/bcbio-docker-build",
+            help="Directory to run docker build in")
+        parser.add_argument(
+            "-q", "--quiet", dest="verbose", action="store_false",
+            default=True, help="Quiet output when running Ansible playbooks")
+        parser.set_defaults(func=self.run)
+
+    def process(self):
+        """Run the command with the received information."""
+        return docker_devel.run_docker_build(self.args)
+
+
+class BiodataUpload(base.BaseCommand):
+
+    """Upload pre-prepared biological data to cache."""
+
+    def setup(self):
+        """Extend the parser configuration in order to expose this command."""
+        parser = self._main_parser.add_parser(
+            "biodata",
+            help="Upload pre-prepared biological data to cache")
+        parser.add_argument(
+            "--prepped",
+            help=("Start with an existing set of cached data to "
+                  "output directory."))
+        parser.add_argument(
+            "--genomes", help="Genomes to download",
+            action="append", default=[],
+            choices=["GRCh37", "hg19", "mm10", "mm9", "rn5", "canFam3", "dm3",
+                     "Zv9", "phix", "sacCer3", "xenTro3", "TAIR10",
+                     "WBcel235"])
+        parser.add_argument(
+            "--aligners", help="Aligner indexes to download",
+            action="append", default=[],
+            choices=["bowtie", "bowtie2", "bwa", "novoalign", "star", "ucsc"])
+        parser.add_argument(
+            "-q", "--quiet", dest="verbose", action="store_false",
+            default=True, help="Quiet output when running Ansible playbooks")
+        parser.set_defaults(func=self.run)
+
+    def process(self):
+        """Manage preparation of biodata on a local machine, uploading
+        to S3 in pieces."""
+        return docker_devel.run_biodata_upload(self.args)
+
+
+class SystemUpdate(base.BaseCommand):
+
+    """Update bcbio system file with a given core and memory/core target."""
+
+    def setup(self):
+        """Extend the parser configuration in order to expose this command."""
+        parser = self._main_parser.add_parser(
+            "system",
+            help=("Update bcbio system file with a given core "
+                  "and memory/core target"))
+        parser.add_argument(
+            "cores",
+            help="Target cores to use for multi-core processes")
+        parser.add_argument(
+            "memory",
+            help="Target memory per core, in Mb (1000 = 1Gb)")
+        parser.add_argument(
+            "-q", "--quiet", dest="verbose", action="store_false",
+            default=True, help="Quiet output when running Ansible playbooks")
+        parser.set_defaults(func=self.run)
+
+    def process(self):
+        """Update bcbio_system.yaml file with a given target of cores
+        and memory.
+        """
+        return docker_devel.run_system_update(self.args)
+
+
+class SetupInstall(base.BaseCommand):
+
+    """Run a python setup.py install inside of the current directory."""
+
+    def setup(self):
+        """Extend the parser configuration in order to expose this command."""
+        parser = self._main_parser.add_parser(
+            "setup_install",
+            help=("Run a python setup.py install inside of "
+                  "the current directory"))
+        parser.add_argument(
+            "-i", "--image", help="Image name to write updates to",
+            default=constant.DOCKER_DEFAULT_IMAGE)
+        parser.add_argument(
+            "-q", "--quiet", dest="verbose", action="store_false",
+            default=True, help="Quiet output when running Ansible playbooks")
+        parser.set_defaults(func=self.run)
+
+    def process(self):
+        """Install python code from a bcbio-nextgen development tree
+        inside of docker.
+        """
+        return docker_devel.run_setup_install(self.args)
